backend/pyproject.toml
[project]
name = "synapse-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "litellm>=1.50.0",
    "langchain-litellm>=0.3.0",
    "langgraph>=0.3.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary,pool]>=3.2.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "fsrs>=1.0.0",
    # Phase 3
    "mem0ai>=0.1.0",
    "docling>=2.0.0",
    "pypdf>=4.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
